tambah fitur tambah data dan hapus data

+++ mahasiswa.php
@@ -28,7 +28,7 @@ $qmahasiswas = tampildata($qmahasiswa); // array associative
         <tr>
             <td><a href="index.php">Home</a></td>
             <td><a href="profil.php">Profile</a></td>
-            <td><a href="contact.php">Contact</a></td>
+            <td><a href="kontak.php">Kontak</a></td>
             <td><a href="mahasiswa.php">Data Mahasiswa</a></td>
         </tr>
     </table>
@@ -70,7 +70,7 @@ $qmahasiswas = tampildata($qmahasiswa); // array associative
             <td><img src="Asset/Image/<?= $mhs["foto"] ?>" width="80"></td>
             <td>
                 <a href="editdata.php"><button>Edit</button></a>
-                <a href="hapusdata.php"><button>Hapus</button></a>
+                <a href="hapusdata.php?id=<?= $mhs['id'] ?>" onclick="return confirm('YAKEUNNN?')"><button>Hapus</button></a>
             </td>
         </tr>
